# api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Relay API",
    "version": "3.1.0",
    "description": "Multi-file employee (or any schema-driven) data migration with AI-assisted column mapping, human escalation, idempotent delivery and rollback. Sign in with POST /api/auth/login (cookie session), then send the returned csrfToken as the x-csrf-token header on every non-GET request."
  },
  "servers": [
    {
      "url": "http://127.0.0.1:3001"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "relay_session"
      },
      "csrf": {
        "type": "apiKey",
        "in": "header",
        "name": "x-csrf-token"
      },
      "service": {
        "type": "apiKey",
        "in": "header",
        "name": "x-relay-service",
        "description": "Internal credential held only by the migration service."
      }
    }
  },
  "security": [
    {
      "session": [],
      "csrf": []
    }
  ],
  "paths": {
...
    "/api/presets": {
      "get": {
        "summary": "Ready-made targets with their schemas and sample files",
        "tags": [
          "Migrations"
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Not signed in"
          }
        }
      }
    },
    "/api/settings": {
      "get": {
        "summary": "AI and alert settings (secrets are never returned)",
        "tags": [
          "Settings"
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Not signed in"
          }
        },
        "description": "Administrator only. "
      }
    },
    "/api/settings/ai": {
      "put": {
        "summary": "Save the OpenRouter key (optional after first save), model, confidence threshold and sample-sharing",
        "tags": [
          "Settings"
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Not signed in"
          }
        },
        "description": "Administrator only. ",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "model": {
                    "type": "string"
                  },
                  "shareSamples": {
                    "type": "boolean"
                  },
                  "minConfidence": {
                    "type": "number",
                    "minimum": 0.5,
                    "maximum": 0.99
                  },
                  "apiKey": {
                    "type": "string"
                  }
                },
                "required": [
                  "model",
                  "shareSamples",
                  "minConfidence"
                ]
              }
            }
          }
        }
      }
    },
    "/api/settings/ai/key": {
      "delete": {
        "summary": "Remove the saved key",
        "tags": [
          "Settings"
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Not signed in"
          }
        },
        "description": "Administrator only. "
      }
    },
    "/api/settings/ai/test": {
      "post": {
        "summary": "Make a tiny real mapping request to verify key and model (before or after saving)",
        "tags": [
          "Settings"
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Not signed in"
          }
        },
        "description": "Administrator only. ",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "model": {
                    "type": "string"
                  },
                  "apiKey": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/settings/ai/models": {
      "get": {
        "summary": "Free models currently offered by OpenRouter",
        "tags": [
          "Settings"
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Not signed in"
          }
        },
        "description": "Administrator only. "
      }
    },
    "/api/settings/escalation": {
      "put": {
        "summary": "Set or clear the Slack-compatible escalation webhook ('' clears)",
        "tags": [
          "Settings"
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Not signed in"
          }
        },
        "description": "Administrator only. ",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "webhookUrl": {
                    "type": "string"
                  }
                },
                "required": [
                  "webhookUrl"
                ]
              }
            }
          }
        }
      }
    },
    "/api/sources/preview": {
      "post": {
        "summary": "Parse files and return sheets, row counts, columns and samples without creating anything",
        "tags": [
          "Migrations"
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Not signed in"
          }
        },
        "description": "Administrator only. ",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "files": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "binary"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/runs": {
      "get": {
        "summary": "Migrations you may see (admins: all; consultants: assigned)",
        "tags": [
          "Migrations"
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Not signed in"
          }
        }
      },
      "post": {
        "summary": "Create a migration and start the agent (Generate)",
        "tags": [
          "Migrations"
        ],
        "responses": {
          "202": {
            "description": "Accepted; poll GET /api/runs/{id}"
          },
          "400": {
            "description": "Unreadable file or invalid schema/destination"
          }
        },
        "description": "Administrator only. ",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "assignedTo": {
                    "type": "string",
                    "description": "Consultant user id who receives escalations"
                  },
                  "configuration": {
                    "type": "string",
                    "description": "JSON: {schema, identityField, destination:{kind:'reference'}|{kind:'http',url}}. Omit for the Employee preset + built-in mock API."
                  },
                  "authorization": {
                    "type": "string",
                    "description": "Optional Authorization header for an http destination; stored encrypted."
                  },
                  "files": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "binary"
                    }
                  }
                },
                "required": [
                  "name",
                  "assignedTo",
                  "files"
                ]
              }
            }
          }
        }
      }
    },
    "/api/runs/{id}": {
      "get": {
        "summary": "Full migration state: mappings, records, review cases, escalation, audit events",
        "tags": [
          "Migrations"
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Not signed in"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/api/runs/{id}/summary": {
      "get": {
        "summary": "Counts, AI usage and current escalation",
        "tags": [
          "Migrations"
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Not signed in"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/api/runs/{id}/audit": {
      "get": {
        "summary": "Download the audit export",
        "tags": [
          "Migrations"
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Not signed in"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/api/runs/{id}/payload": {
      "get": {
        "summary": "The exact JSON records that will be / were delivered (blocked while cases are open)",
        "tags": [
          "Migrations"
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Not signed in"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/api/runs/{id}/export": {
      "get": {
        "summary": "Cleaned records as CSV",
        "tags": [
          "Migrations"
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Not signed in"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
...
    "/api/runs/{id}/retry": {
      "post": {
        "summary": "Retry outstanding (failed or unsent) records",
        "tags": [
          "Recovery"
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Not signed in"
          }
        },
        "description": "Administrator only. Same idempotency key per record, so retries never duplicate writes.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
...
  }
}